fix: break verify↔service-datasource build cycle; run driver conformance in dogfood
The prior commit added @objectstack/verify as a devDep of service-datasource,
but verify already depends on service-datasource (via runtime) — turbo flagged
the cycle and the build/typecheck/test jobs failed.

- Move the cross-driver conformance test out of service-datasource into
  packages/dogfood (which already dev-tests through @objectstack/verify and runs
  as the Dogfood Regression Gate); add driver-sql + driver-memory as dogfood
  devDeps. No cycle: drivers don't depend on verify/dogfood.
- Revert the service-datasource test + devDep.
- checkReadCoercion now reads back its own row by id (robust when the probe
  object already holds rows); the memory-driver test uses `persistence: false`
  so the probe never leaks to a shared on-disk snapshot across suites.

Verified: `turbo run build --filter=./packages/*` 60/60 (no cycle); full dogfood
suite 191 passed / 39 files (incl. the new conformance test: both drivers → [],
non-coercing stub → 3 problems); verify build incl. DTS clean; eslint clean.
