+++ packages/services/service-datasource/package.json
@@ -35,7 +35,6 @@
     "@objectstack/driver-sql": "workspace:*",
     "@objectstack/driver-sqlite-wasm": "workspace:*",
     "@objectstack/plugin-hono-server": "workspace:*",
-    "@objectstack/verify": "workspace:*",
     "@types/node": "^26.1.0",
     "tsup": "^8.5.1",
     "typescript": "^6.0.3",
